week10/ProjetFinal/generator/requirements.txt
Django==4.0.3
Pillow
django-qr-code
